Organize routes and lines components into own directories
