+++ src/components/routes-and-lines/line-details/DirectionBadge.tsx
@@ -1,4 +1,4 @@
-import { RouteDirectionEnum } from '../../generated/graphql';
+import { RouteDirectionEnum } from '../../../generated/graphql';
 
 type Props = {
   direction: RouteDirectionEnum;
