+++ src/components/routes-and-lines/line-details/RouteStopsTable.tsx
@@ -2,8 +2,8 @@ import { Switch as HuiSwitch } from '@headlessui/react';
 import orderBy from 'lodash/orderBy';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { RouteRoute } from '../../generated/graphql';
-import { Switch, SwitchLabel } from '../../uiComponents';
+import { RouteRoute } from '../../../generated/graphql';
+import { Switch, SwitchLabel } from '../../../uiComponents';
 import { RouteStopsSection } from './RouteStopsSection';
 
 interface Props {
