+++ src/components/routes-and-lines/main/RoutesAndLinesPage.tsx
@@ -1,13 +1,13 @@
 import React, { useContext } from 'react';
 import { useTranslation } from 'react-i18next';
-import { MapEditorContext } from '../context/MapEditor';
-import { useAppDispatch, useMapUrlQuery } from '../hooks';
-import { Container, Row } from '../layoutComponents';
-import { setIsModalMapOpenAction } from '../redux';
-import { Path, routeDetails } from '../router/routeDetails';
-import { SimpleButton } from '../uiComponents';
-import { RoutesAndLinesLists } from './routes-and-lines/RoutesAndLinesLists';
-import { SearchContainer } from './routes-and-lines/search/conditions/SearchContainer';
+import { MapEditorContext } from '../../../context/MapEditor';
+import { useAppDispatch, useMapUrlQuery } from '../../../hooks';
+import { Container, Row } from '../../../layoutComponents';
+import { setIsModalMapOpenAction } from '../../../redux';
+import { Path, routeDetails } from '../../../router/routeDetails';
+import { SimpleButton } from '../../../uiComponents';
+import { SearchContainer } from '../search/conditions/SearchContainer';
+import { RoutesAndLinesLists } from './RoutesAndLinesLists';
 
 export const RoutesAndLinesPage = (): JSX.Element => {
   const { t } = useTranslation();
