+++ src/components/routes-and-lines/main/RoutesTable.spec.tsx
@@ -4,9 +4,12 @@ import {
   RouteRoute,
   useListChangingRoutesQuery,
   useListOwnLinesQuery,
-} from '../generated/graphql';
-import { mapListChangingRoutesResult, mapListOwnLinesResult } from '../graphql';
-import { render } from '../utils/test-utils';
+} from '../../../generated/graphql';
+import {
+  mapListChangingRoutesResult,
+  mapListOwnLinesResult,
+} from '../../../graphql';
+import { render } from '../../../utils/test-utils';
 import { LineTableRow } from './LineTableRow';
 import { RoutesTable } from './RoutesTable';
 import { RoutesTableRow } from './RoutesTableRow';
